TraceGenerator, docs and clean up
* introduced `ElementGenerator` as base class for any generator that isn't an `Importer`
* implement `TraceGenerator` repr. a log by the xes-standard definition
* tests for `XESImporter` extracting meta data
* tests for `TraceGenerator`
* lots of documentation
* moved `Event` and `Trace` to base.py as it makes more sense there
* merged events.py and traces.py to generators.py
